migrate pstart to uint64

+++ internal/store/proxmox/types.go
@@ -17,7 +17,7 @@ type Task struct {
 	WID        string `json:"id"`
 	Node       string `json:"node"`
 	PID        int    `json:"pid"`
-	PStart     int    `json:"pstart"`
+	PStart     uint64 `json:"pstart"`
 	StartTime  int64  `json:"starttime"`
 	EndTime    int64  `json:"endtime"`
 	TaskId     string `json:"task_id"`
